--- src/main/kotlin/practice/TuplePractice.kt
@@ -1,4 +1,0 @@
-package practice
-
-class TuplePractice {
-}
