+++ src/main/kotlin/tuple/Tuple.kt
@@ -1,3 +1,5 @@
+package tuple
+
 import kotlin.math.abs
 import kotlin.math.sqrt
