+++ src/test/kotlin/tuple/TupleTest.kt
@@ -1,3 +1,5 @@
+package tuple
+
 import assertk.all
 import assertk.assertThat
 import assertk.assertions.*
