+++ src/test/kotlin/tuple/practice/TextVisualizerTest.kt
@@ -1,10 +1,9 @@
-package practice
+package tuple.practice
 
-import Point
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import org.junit.jupiter.api.Test
-import point
+import tuple.point
 
 class TextVisualizerTest {
 
@@ -63,25 +62,6 @@ class TextVisualizerTest {
             |#*
         """.trimMargin()
         )
-    }
-
-    private fun visualize(
-        size: Int,
-        default: Char = '#',
-        pointChar: Char = '*',
-        points: List<Point> = emptyList(),
-    ): String {
-        return (0 until size).joinToString(separator = "\n") { y ->
-            (0 until size).joinToString(separator = "") { x ->
-                val point = points.find { it.x.toInt() == x && it.y.toInt() == y }
-                if (point != null) {
-                    pointChar.toString()
-                } else {
-                    default.toString()
-                }
-
-            }
-        }
     }
 
 }
